requirements.txt
llama-cpp-python
faiss-cpu
sentence-transformers
networkx
matplotlib
huggingface_hub
tqdm
colorama
